backend/requirements.txt
fastapi
uvicorn
supabase
psycopg2-binary
pandas
openpyxl
python-dotenv
google-generativeai
google-genai
python-multipart
httpx
pydantic